Make sure all mocks live under a mocks.go file, for consistency.
