+++ internal/apptracker/sentry/sentry_tracker_test.go
@@ -3,79 +3,11 @@ package sentry
 import (
 	"errors"
 	"testing"
-	"time"
 
 	"github.com/getsentry/sentry-go"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
-
-// MockSentry is a mock struct to capture function calls
-type MockSentry struct {
-	mock.Mock
-}
-
-func (m *MockSentry) CaptureMessage(message string) *sentry.EventID {
-	args := m.Called(message)
-	return args.Get(0).(*sentry.EventID)
-}
-
-func (m *MockSentry) CaptureException(exception error) *sentry.EventID {
-	args := m.Called(exception)
-	return args.Get(0).(*sentry.EventID)
-}
-
-func (m *MockSentry) Init(options sentry.ClientOptions) error {
-	args := m.Called(options)
-	return args.Error(0)
-}
-
-func (m *MockSentry) Flush(timeout time.Duration) bool {
-	m.Called(timeout)
-	return true
-}
-
-func (m *MockSentry) Recover() *sentry.EventID {
-	args := m.Called()
-	return args.Get(0).(*sentry.EventID)
-}
-
-const testSentryDSN = "https://username@password.test.com/some-id"
-
-func setupMockSentry(t *testing.T) *MockSentry {
-	t.Helper()
-
-	mockSentry := &MockSentry{}
-	mockSentry.
-		On("Flush", mock.Anything).Return(true).Once().
-		On("Recover").Return((*sentry.EventID)(nil)).Once()
-
-	// Save the original functions
-	originalInitFunc := InitFunc
-	originalFlushFunc := FlushFunc
-	originalRecoverFunc := RecoverFunc
-	originalCaptureMessageFunc := captureMessageFunc
-	originalCaptureExceptionFunc := captureExceptionFunc
-
-	// Set the mock functions
-	InitFunc = mockSentry.Init
-	FlushFunc = mockSentry.Flush
-	RecoverFunc = mockSentry.Recover
-	captureMessageFunc = mockSentry.CaptureMessage
-	captureExceptionFunc = mockSentry.CaptureException
-
-	// Restore the original functions after the test
-	t.Cleanup(func() {
-		InitFunc = originalInitFunc
-		FlushFunc = originalFlushFunc
-		RecoverFunc = originalRecoverFunc
-		captureMessageFunc = originalCaptureMessageFunc
-		captureExceptionFunc = originalCaptureExceptionFunc
-		mockSentry.AssertExpectations(t)
-	})
-
-	return mockSentry
-}
 
 func TestSentryTracker_CaptureMessage(t *testing.T) {
 	mockSentry := setupMockSentry(t)
@@ -83,7 +15,7 @@ func TestSentryTracker_CaptureMessage(t *testing.T) {
 		On("Init", mock.Anything).Return(nil).Once().
 		On("CaptureMessage", "Test message").Return((*sentry.EventID)(nil)).Once()
 
-	tracker, err := NewSentryTracker(testSentryDSN, "test", 5)
+	tracker, err := NewSentryTracker("dsn", "test-env", 5)
 	require.NoError(t, err)
 	require.NotNil(t, tracker)
 
@@ -97,7 +29,7 @@ func TestSentryTracker_CaptureException(t *testing.T) {
 		On("Init", mock.Anything).Return(nil).Once().
 		On("CaptureException", testError).Return((*sentry.EventID)(nil)).Once()
 
-	tracker, err := NewSentryTracker(testSentryDSN, "test", 5)
+	tracker, err := NewSentryTracker("dsn", "test-env", 5)
 	require.NoError(t, err)
 	require.NotNil(t, tracker)
