+++ internal/services/mocks.go
@@ -5,6 +5,8 @@ import (
 
 	"github.com/stretchr/testify/mock"
 
+	"github.com/stellar/go/txnbuild"
+
 	"github.com/stellar/wallet-backend/internal/entities"
 )
 
@@ -51,4 +53,20 @@ func (r *RPCServiceMock) GetLedgerEntries(keys []string) (entities.RPCGetLedgerE
 func (r *RPCServiceMock) GetAccountLedgerSequence(address string) (int64, error) {
 	args := r.Called(address)
 	return args.Get(0).(int64), args.Error(1)
+}
+
+type AccountSponsorshipServiceMock struct {
+	mock.Mock
+}
+
+var _ AccountSponsorshipService = (*AccountSponsorshipServiceMock)(nil)
+
+func (s *AccountSponsorshipServiceMock) SponsorAccountCreationTransaction(ctx context.Context, accountToSponsor string, signers []entities.Signer, assets []entities.Asset) (string, string, error) {
+	args := s.Called(ctx, accountToSponsor, signers, assets)
+	return args.String(0), args.String(1), args.Error(2)
+}
+
+func (s *AccountSponsorshipServiceMock) WrapTransaction(ctx context.Context, tx *txnbuild.Transaction) (string, string, error) {
+	args := s.Called(ctx, tx)
+	return args.String(0), args.String(1), args.Error(2)
 }
